requirements.txt
flask==2.3.3
flask_sqlalchemy==3.1.1
pyjwt==2.8.0
firebase-admin==6.4.0
werkzeug==2.3.7
gunicorn